+++ crates/selector4nix-core/src/infrastructure/mod.rs
@@ -2,7 +2,6 @@ pub mod cdn;
 pub mod config;
 pub mod dns;
 pub mod endpoint;
-pub mod fastly;
 pub mod metric;
 pub mod provider;
 pub mod repository;