+++ crates/selector4nix-core/src/infrastructure/provider/nar_info_provider.rs
@@ -196,7 +196,6 @@ mod tests {
     use selector4nix_streaming::throttler::{PerHostHttpThrottler, ThrottlingOptions};
 
     use super::*;
-    use crate::infrastructure::dns::doh_resolver::DohResolver;
     use crate::infrastructure::provider::{
         EndpointClientPool, EndpointProbingProvider, SniProxySourceProvider,
     };
@@ -223,10 +222,7 @@ mod tests {
             Url::new(&format!("http://cache.nixos.org:{port}")).unwrap(),
             pool,
             probing,
-            Arc::new(DohResolver::new()),
             user_candidates,
-            false,
-            Vec::new(),
             Vec::new(),
             Arc::new(SniProxySourceProvider::new()),
             None,